+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 
 * Changed development autoreload mechanism to not interfere with Django's
   development server autoreload.
+* Bug fix for case when invalid mode is specified in template tag.
 
 0.9 (2018-09-10)
 ++++++++++++++++